# unofficial-tractive-rest-api/swagger-spec.json
{"openapi":"3.0.0","paths":{"/location/{trackerId}":{"get":{"operationId":"LocationController_getTrackerLocation","parameters":[{"name":"trackerId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Location"]}},"/auth":{"get":{"operationId":"AuthController_authenticate","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/is-authenticated":{"get":{"operationId":"AuthController_isUserAuthenticated","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/hardware/{trackerId}":{"get":{"operationId":"HardwareController_getHardwareReport","parameters":[{"name":"trackerId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Hardware"]}},"/hardware/battery/{trackerId}":{"get":{"operationId":"HardwareController_getBattery","parameters":[{"name":"trackerId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Hardware"]}},"/account/info":{"get":{"operationId":"AccountController_getAccountInfo","parameters":[],"responses":{"200":{"description":""}},"tags":["Account"]}},"/account/subscriptions":{"get":{"operationId":"AccountController_getAccountSubscriptions","parameters":[],"responses":{"200":{"description":""}},"tags":["Account"]}},"/account/subscriptions/{subscriptionID}":{"get":{"operationId":"AccountController_getAccountSubscription","parameters":[{"name":"subscriptionID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Account"]}},"/account/shares":{"get":{"operationId":"AccountController_getAccountShares","parameters":[],"responses":{"200":{"description":""}},"tags":["Account"]}},"/pet":{"get":{"operationId":"PetController_getPets","parameters":[],"responses":{"200":{"description":""}},"tags":["Pet"]}},"/pet/{petID}":{"get":{"operationId":"PetController_getPet","parameters":[{"name":"petID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Pet"]}},"/pet/{petID}/health":{"get":{"operationId":"PetController_getPetHealth","parameters":[{"name":"petID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Pet"]}},"/command/{trackerID}/live/on":{"post":{"operationId":"CommandController_liveOn","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Command"]}},"/command/{trackerID}/live/off":{"post":{"operationId":"CommandController_liveOff","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Command"]}},"/command/{trackerID}/led/on":{"post":{"operationId":"CommandController_ledOn","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Command"]}},"/command/{trackerID}/led/off":{"post":{"operationId":"CommandController_ledOff","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Command"]}},"/command/{trackerID}/buzzer/on":{"post":{"operationId":"CommandController_buzzerOn","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Command"]}},"/command/{trackerID}/buzzer/off":{"post":{"operationId":"CommandController_buzzerOff","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Command"]}},"/tracker":{"get":{"operationId":"TrackerController_getAllTrackers","parameters":[],"responses":{"200":{"description":""}},"tags":["Tracker"]}},"/tracker/{trackerID}":{"get":{"operationId":"TrackerController_getTracker","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tracker"]}},"/tracker/{trackerID}/history":{"get":{"operationId":"TrackerController_getTrackerHistory","parameters":[{"name":"trackerID","required":true,"in":"path","schema":{"type":"string"}},{"name":"from","required":true,"in":"query","schema":{"type":"string"}},{"name":"to","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tracker"]}}},"info":{"title":"Tractive REST API","description":"\nTractive is a company that specializes in GPS tracking devices for pets, primarily dogs and cats. These devices allow\nowners to keep tabs on the location of their pets in real-time via a mobile app. The Tractive GPS tracker can be easily\nattached to the pet's collar and uses a combination of GPS, Wi-Fi, and cellular technology to provide accurate location\ndata. The service usually requires a subscription fee for access to the GPS tracking features.\n\nIn addition to tracking, some Tractive devices offer additional features such as activity monitoring, which gives you\ninsights into your pet's daily activities and behaviors. This can be useful for pet owners looking to monitor their\npet's health and well-being closely.\n\nSince tractive has no official, documented public API, this will be a wrapper around their REST API.\n","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{}}}